Update the related Lua bindings for experimental
