--- cocos/scripting/lua-bindings/auto/lua_cocos2dx_experimental_video_auto.hpp
@@ -1,34 +1,0 @@
-#include "base/ccConfig.h"
-#if (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID || CC_TARGET_PLATFORM == CC_PLATFORM_IOS)
-#ifndef __cocos2dx_experimental_video_h__
-#define __cocos2dx_experimental_video_h__
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-#include "tolua++.h"
-#ifdef __cplusplus
-}
-#endif
-
-int register_all_cocos2dx_experimental_video(lua_State* tolua_S);
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-#endif // __cocos2dx_experimental_video_h__
-#endif //#if (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID || CC_TARGET_PLATFORM == CC_PLATFORM_IOS)
