+++ tools/tolua/cocos2dx.ini
@@ -132,7 +132,8 @@ skip = Node::[setGLServerState description getUserObject .*UserData getGLServerS
         Animation3DCache::[*],
         Sprite3DMaterialCache::[*],
         Bone3D::[*],
-        FastTMXLayer::[(g|s)etTiles getTileGIDAt]
+        FastTMXLayer::[*],
+        FastTMXTiledMap::[*]
 
 rename_functions = SpriteFrameCache::[addSpriteFramesWithFile=addSpriteFrames getSpriteFrameByName=getSpriteFrame],
     ProgressTimer::[setReverseProgress=setReverseDirection],
